[thread:task] Ref | New queue API
The previous implementation used to rely on static methods and variables, w/o any encapsulation apart from the one provided by modules. The current version, essentially, offers the same functionality, but adheres to "objectish-oriented" paradign to a slightly higher extent.

The previous implementation is still retained. It will be replaced w/ the new one on the next stage.